Add a cache for `cuobjdump` output when running tests (#2)
**tl;dr: This PR makes the test suite run ~17x faster**

This PR introduces a cache for the output of the external/real
`cuobjdump` process since it takes the majority of test execution time.

Locally, the test suite with no caching runs in ~300 seconds. With an
uncompressed cache, it runs in about 5 seconds. Unfortunately, an
uncompressed cache of `cuobjdump` output is ~7gb. This may not be
practical for us to store in CI along with cargo caches because GitHub
Actions has a 10gb limit.

Compressing each cached item with zstd brings the cache size down to
~600mb and test execution to ~18 seconds.

### Test Plan

Ran CI twice and the `cargo test` run in the second execution was 13 to
14 times faster